+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - *WHS.init class -> WHS.World class*
 - #63: Fixed JSDoc.
 - #60: Update lights for Three.js r75
+- *src/Skybox/Skybox.js -> src/Components/Skybox.js*
 
 **v0.0.7**
 - Fixed #16 "object.assign() chrome bug."